+++ src/cli.rs
@@ -1,4 +1,3 @@
-use crate::log::Level;
 use clap::{Args, Parser};
 use std::path::{Path, PathBuf};
 
@@ -8,13 +7,6 @@ pub struct CliArgs {
 	/// Path to the main configuration directory
 	#[arg(short, long, value_name = "DIR", default_value = get_default_config_dir().into_os_string())]
 	pub config: PathBuf,
-
-	/// Specify the log level
-	#[arg(long, value_name = "LEVEL", value_enum, default_value_t = crate::DEFAULT_LOG_LEVEL)]
-	pub log_level: Level,
-
-	#[command(flatten)]
-	pub log: Log,
 
 	/// Runs in the foreground
 	#[arg(short, long, default_value_t = false)]
@@ -26,18 +18,6 @@ pub struct CliArgs {
 	/// Add a root certificate to the trust store (can be set multiple times)
 	#[arg(long, value_name = "FILE")]
 	pub root_cert: Vec<PathBuf>,
-}
-
-#[derive(Args, Debug)]
-#[group(multiple = false)]
-pub struct Log {
-	/// Sends log messages via syslog
-	#[arg(long)]
-	pub log_syslog: bool,
-
-	/// Prints log messages to the standard error output
-	#[arg(long)]
-	pub log_stderr: bool,
 }
 
 #[derive(Args, Debug)]
@@ -96,9 +76,6 @@ mod tests {
 		let args: &[&str] = &[];
 		let pa = CliArgs::try_parse_from(args).unwrap();
 		assert_eq!(pa.config, get_default_config_dir());
-		assert_eq!(pa.log_level, Level::Warn);
-		assert_eq!(pa.log.log_syslog, false);
-		assert_eq!(pa.log.log_stderr, false);
 		assert_eq!(pa.foreground, false);
 		assert_eq!(pa.pid.pid_file, get_default_pid_file());
 		assert_eq!(pa.pid.no_pid_file, false);
@@ -115,9 +92,6 @@ mod tests {
 			"acmed",
 			"--config",
 			"/tmp/test.toml",
-			"--log-level",
-			"debug",
-			"--log-syslog",
 			"--foreground",
 			"--pid-file",
 			"/tmp/debug/acmed.pid",
@@ -130,9 +104,6 @@ mod tests {
 		];
 		let pa = CliArgs::try_parse_from(argv).unwrap();
 		assert_eq!(pa.config, PathBuf::from("/tmp/test.toml"));
-		assert_eq!(pa.log_level, Level::Debug);
-		assert_eq!(pa.log.log_syslog, true);
-		assert_eq!(pa.log.log_stderr, false);
 		assert_eq!(pa.foreground, true);
 		assert_eq!(
 			pa.pid.get_pid_file(),
@@ -154,9 +125,6 @@ mod tests {
 			"acmed",
 			"--config",
 			"/tmp/test.toml",
-			"--log-level",
-			"debug",
-			"--log-stderr",
 			"--foreground",
 			"--no-pid-file",
 			"--root-cert",
@@ -168,9 +136,6 @@ mod tests {
 		];
 		let pa = CliArgs::try_parse_from(argv).unwrap();
 		assert_eq!(pa.config, PathBuf::from("/tmp/test.toml"));
-		assert_eq!(pa.log_level, Level::Debug);
-		assert_eq!(pa.log.log_syslog, false);
-		assert_eq!(pa.log.log_stderr, true);
 		assert_eq!(pa.foreground, true);
 		assert_eq!(pa.pid.get_pid_file(), None);
 		assert_eq!(
@@ -189,9 +154,6 @@ mod tests {
 			"acmed",
 			"-c",
 			"/tmp/test.toml",
-			"--log-level",
-			"debug",
-			"--log-syslog",
 			"-f",
 			"--pid-file",
 			"/tmp/debug/acmed.pid",
@@ -204,9 +166,6 @@ mod tests {
 		];
 		let pa = CliArgs::try_parse_from(argv).unwrap();
 		assert_eq!(pa.config, PathBuf::from("/tmp/test.toml"));
-		assert_eq!(pa.log_level, Level::Debug);
-		assert_eq!(pa.log.log_syslog, true);
-		assert_eq!(pa.log.log_stderr, false);
 		assert_eq!(pa.foreground, true);
 		assert_eq!(
 			pa.pid.get_pid_file(),
@@ -228,9 +187,6 @@ mod tests {
 			"acmed",
 			"-c",
 			"/tmp/test.toml",
-			"--log-level",
-			"debug",
-			"--log-stderr",
 			"-f",
 			"--no-pid-file",
 			"--root-cert",
@@ -242,9 +198,6 @@ mod tests {
 		];
 		let pa = CliArgs::try_parse_from(argv).unwrap();
 		assert_eq!(pa.config, PathBuf::from("/tmp/test.toml"));
-		assert_eq!(pa.log_level, Level::Debug);
-		assert_eq!(pa.log.log_syslog, false);
-		assert_eq!(pa.log.log_stderr, true);
 		assert_eq!(pa.foreground, true);
 		assert_eq!(pa.pid.get_pid_file(), None);
 		assert_eq!(
@@ -255,13 +208,6 @@ mod tests {
 				PathBuf::from("/tmp/certs/root_03.pem")
 			]
 		);
-	}
-
-	#[test]
-	fn err_log_output() {
-		let argv: &[&str] = &["acmed", "--log-stderr", "--log-syslog"];
-		let pa = CliArgs::try_parse_from(argv);
-		assert!(pa.is_err());
 	}
 
 	#[test]